scripts/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
PyPDF2>=3.0.0
psycopg2-binary>=2.9.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
python-docx>=1.1.0
